User words file is unexpectedly removed from disk
PDFOC-93

+++ pdfocr-tesseract4/src/main/java/com/itextpdf/pdfocr/tesseract4/Tesseract4ExecutableOcrEngine.java
@@ -183,7 +183,8 @@ public class Tesseract4ExecutableOcrEngine extends AbstractTesseract4OcrEngine {
             }
             try {
                 if (getTesseract4OcrEngineProperties()
-                        .getPathToUserWordsFile() != null) {
+                        .getPathToUserWordsFile() != null
+                        && getTesseract4OcrEngineProperties().isUserWordsFileTemporary()) {
                     TesseractHelper.deleteFile(
                             getTesseract4OcrEngineProperties()
                                     .getPathToUserWordsFile());
